@@ -1,38 +1,46 @@
 {
-  "name": "my-starter-ts-npm-cli-and-lib",
-  "version": "0.1.3",
-  "description": "Typescript で CLI とライブラリの npm パッケージを作成するためのスターター。",
+  "name": "@hankei6km/remark-numbers",
+  "version": "0.1.0",
+  "description": "remark plugin to assign serial numbers",
+  "author": "hankei6km <hankei6km@gmail.com> (https://github.com/hankei6km)",
   "license": "MIT",
-  "keywords": [],
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/hankei6km/remark-numbers.git"
+  },
+  "bugs": {
+    "url": "https://github.com/hankei6km/remark-numbers/issues"
+  },
+  "keywords": [
+    "remnrk"
+  ],
   "main": "dist/index.js",
+  "exports": "./dist/index.js",
+  "type": "module",
   "types": "dist/index.d.ts",
   "files": [
     "dist"
   ],
-  "bin": {
-    "count": "dist/main.js"
-  },
   "scripts": {
-    "start": "tsc && node dist/main.js",
-    "start:watch": "nodemon --watch src --ext ts --exec ts-node src/main.ts",
-    "build": "tsc",
-    "test": "jest",
-    "clean": "rm dist/*",
+    "start": "npm run build && node dist/main.js",
+    "build": "npm run clean && tsc && rimraf dist/test && mv dist/src/* dist/ && rimraf dist/src",
+    "test": "node --experimental-vm-modules node_modules/.bin/jest",
+    "clean": "rimraf \"dist/*\"",
     "upgrade-interactive": "npm-check --update",
     "csb:test": "npm test -- --runInBand --watchAll"
   },
   "dependencies": {
-    "yargs": "^17.2.1"
+    "yargs": "^17.3.1"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.1",
-    "@types/yargs": "^17.0.4",
-    "jest": "^27.3.1",
-    "nodemon": "^2.0.14",
-    "npm-check": "^5.9.2",
-    "ts-jest": "^27.0.7",
-    "ts-node": "^10.3.0",
-    "typescript": "^4.4.4"
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.6",
+    "@types/yargs": "^17.0.8",
+    "jest": "^27.4.5",
+    "nodemon": "^2.0.15",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^27.1.2",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4"
   }
 }
